signalmanager: Move some functions to utils
